+++ docs/changes.rst
@@ -17,6 +17,7 @@ Unreleased.
  - Added :meth:`Image.affine() <wand.image.BaseImage.affine>` method.
  - Added :meth:`Image.bilateral_blur() <wand.image.BaseImage.bilateral_blur>` method.
  - Added :meth:`Image.evaluate_images() <wand.image.BaseImage.evaluate_images>` method.
+ - Added :meth:`Image.floodfill() <wand.image.BaseImage.floodfill>` method.
  - Added :meth:`Image.morph() <wand.image.Image.morph>` method.
  - Fixed :meth:`Image.quantize() <wand.image.BaseImage.quantize>` behavior by switching
    default value of ``colorspace_type`` from :const:`None` to ``"undefined"``. [:issue:`644`]